+++ packages/cli/src/lionweb-cli.ts
@@ -1,3 +1,6 @@
+#!/usr/bin/env node
+
+
 import {argv} from "process"
 
 import {extractFromSerialization} from "./serialization-extractor.js"
@@ -14,7 +17,7 @@ const main = (args: string[])=> {
 
     if (args.length <= 2) {
         console.log(
-            `lionweb-cli is a LionWeb utility around LionWeb-TypeScript
+`lionweb-cli is a LionWeb utility around LionWeb-TypeScript
 
 Usage: $ npx @lionweb/cli <command> <arguments>
 
@@ -45,7 +48,8 @@ ${commands.map((command) => `    ${command}\n`).join(``)}
             if (commandArgs.length === 0) {
                 console.log(
 `The ${EXTRACT_COMMAND} command extracts the following from a serialization chunk in the form of files: a sorted JSON, and a shortened JSON.
-If the chunk is the serialization of a LionCore Language/M2, then a textual rendering is already output.`
+If the chunk is the serialization of a LionCore Language/M2, then a textual rendering is already output.
+(See the README.md for more information.)`
                 )
             } else {
                 commandArgs.forEach(extractFromSerialization)
@@ -64,7 +68,7 @@ If the chunk is the serialization of a LionCore Language/M2, then a textual rend
             return
         }
 
-        // TODO  schema, Ecore import
+        // TODO  generate schema, import Ecore
 
         default: {
             console.error(`command "${command}" is not recognized`)