+++ packages/core/src/m3/builtins.ts
@@ -1,14 +1,13 @@
 import {LanguageFactory} from "./factory.js"
 import {lioncoreBuiltinsQName, lioncoreQNameSeparator, Property} from "./types.js"
 import {qualifiedNameBasedKeyGenerator} from "./key-generation.js"
-import {checkAll} from "../id-generation.js"
 import {currentReleaseVersion} from "../version.js"
 
 
 const factory = new LanguageFactory(
     lioncoreBuiltinsQName,
     currentReleaseVersion,
-    checkAll((qualifiedName) => qualifiedName!),
+    (qualifiedName) => qualifiedName!,
     qualifiedNameBasedKeyGenerator(lioncoreQNameSeparator)
 )
 
@@ -27,7 +26,7 @@ const jsonDatatype = factory.primitiveType("JSON")
 const node = factory.concept("Node", true)
 
 
-const inamed = factory.conceptInterface("INamed")
+const inamed = factory.interface("INamed")
 
 const inamed_name = factory.property(inamed, "name")
     .ofType(stringDatatype)